Update mat_batch_tools.py
Added a new operator, "Find Active Face Texture" - searches for the diffuse texture assigned to the active face, and then loads it in the Image Editor. Can be found via search menu in the Image Editor

Added new option to the "Convert to Lightmapped Material" operator. It now supports using vertex colors as the lightmap, instead of an image texture. The vertex color layer in the mesh object(s) must be named "light"