Update mat_batch_tools.py
Added two new material templates "Mirror UV" and "Unmirror UV", for instantly adding a mirror / ping pong effect to the UV Map in all materials in all selected objects. The Unmirror template can be used to remove the effect afterward, with the caveat that it will only work if the effect was originally applied by this addon itself. (ie. The Unmirror template won't work if the ping pong effect was added manually by the user or was created during import of the model into Blender)