Update __init__.py
Fixed small bug in the new Isolate by Material Trait operator. If the operator was used on a mesh object that had already been isolated and separated before, it was re-isolating it into another object, and then leaving an empty object (with no polygons) as well. The object is now correctly skipped if it detects that *all* of the mesh's faces are about to be isolated.