Update __init__.py
Rewrote the 'Copy Dfiuse Texture to Material Name' operator so that it can handle materials with multiple diffuse textures. If a material contains more than 1 texture, the operator will now append all of them to the material name. If a material with the new name already exists,, the operator will assign that existing material to that material slot, so that duplicate materials aren't created. The operator also compares the node count in materials when checking for an existing one with the same name, and if the node count is different, the amount of nodes will be appended to the material name to keep it as a separate material and avoid replacing a material with the same name but different node count.